feat: Refactor product details page and enhance metadata generation for improved SEO
